Added parallelization
