+++ examples/rag-docling-weaviate-agent/requirements.txt
@@ -1,2 +1,2 @@
 docling>=2.7.0
-torch>=2.2.0 
+transformers>=4.50.3
